# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXA Payment Crypto API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  id: number;
  ownerName: string;
  totalValueUsd: number;
  totalValueChange24h: number;
  totalValueChangePct24h: number;
  createdAt: string;
}

export interface WalletAddress {
  symbol: string;
  name: string;
  address: string;
  network: string;
  /** @nullable */
  qrCode?: string | null;
}

export interface Asset {
  id: number;
  symbol: string;
  name: string;
  balance: number;
  valueUsd: number;
  priceUsd: number;
  change24h: number;
  changePct24h: number;
  logoColor: string;
  /** @nullable */
  allocationPct?: number | null;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  send: 'send',
  receive: 'receive',
  swap: 'swap',
  nfc_pay: 'nfc_pay',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  asset: string;
  amount: number;
  valueUsd: number;
  toAddress: string;
  fromAddress: string;
  status: TransactionStatus;
  createdAt: string;
  hash: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  merchantName?: string | null;
  /** @nullable */
  cardId?: number | null;
}

export interface TransactionInput {
  asset: string;
  amount: number;
  toAddress: string;
  note?: string;
}

export interface TransactionSummary {
  totalSent: number;
  totalReceived: number;
  totalNfcPay: number;
  totalSwap: number;
  countSent: number;
  countReceived: number;
  countNfcPay: number;
}

export type CryptoCardStatus = typeof CryptoCardStatus[keyof typeof CryptoCardStatus];


export const CryptoCardStatus = {
  active: 'active',
  frozen: 'frozen',
  cancelled: 'cancelled',
} as const;

export type CryptoCardCardType = typeof CryptoCardCardType[keyof typeof CryptoCardCardType];


export const CryptoCardCardType = {
  virtual: 'virtual',
  physical: 'physical',
} as const;

export interface CryptoCard {
  id: number;
  cardNumber: string;
  cardholderName: string;
  expiryMonth: number;
  expiryYear: number;
  cvv: string;
  asset: string;
  spendingLimit: number;
  spentAmount: number;
  status: CryptoCardStatus;
  cardType: CryptoCardCardType;
  createdAt: string;
  nfcEnabled?: boolean;
}

export type CardInputCardType = typeof CardInputCardType[keyof typeof CardInputCardType];


export const CardInputCardType = {
  virtual: 'virtual',
  physical: 'physical',
} as const;

export interface CardInput {
  cardholderName: string;
  asset: string;
  spendingLimit: number;
  cardType: CardInputCardType;
  nfcEnabled?: boolean;
}

export type CardUpdateStatus = typeof CardUpdateStatus[keyof typeof CardUpdateStatus];


export const CardUpdateStatus = {
  active: 'active',
  frozen: 'frozen',
  cancelled: 'cancelled',
} as const;

export interface CardUpdate {
  status?: CardUpdateStatus;
  spendingLimit?: number;
  nfcEnabled?: boolean;
}

export type NfcSessionStatus = typeof NfcSessionStatus[keyof typeof NfcSessionStatus];


export const NfcSessionStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  expired: 'expired',
} as const;

export interface NfcSession {
  sessionToken: string;
  expiresAt: string;
  merchantName: string;
  amount: number;
  asset: string;
  status: NfcSessionStatus;
}

export interface NfcSessionInput {
  amount: number;
  asset: string;
  merchantName: string;
  cardId?: number;
}

export interface NfcConfirmInput {
  sessionToken: string;
  cardId?: number;
}

export interface MarketPrice {
  symbol: string;
  name: string;
  priceUsd: number;
  change24h: number;
  changePct24h: number;
  marketCapUsd: number;
  volume24hUsd: number;
  logoColor: string;
}

export interface ChartPoint {
  timestamp: string;
  price: number;
}

export interface DashboardSummary {
  totalValueUsd: number;
  change24h: number;
  changePct24h: number;
  totalAssets: number;
  activeCards: number;
  nfcPaymentsThisMonth: number;
  topGainer: string;
  topLoser: string;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  send: 'send',
  receive: 'receive',
  swap: 'swap',
  nfc_pay: 'nfc_pay',
  card_issued: 'card_issued',
  price_alert: 'price_alert',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  subtitle: string;
  /** @nullable */
  valueUsd: number | null;
  timestamp: string;
  icon: string;
}

export type ListTransactionsParams = {
type?: ListTransactionsType;
asset?: string;
limit?: number;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  send: 'send',
  receive: 'receive',
  swap: 'swap',
  nfc_pay: 'nfc_pay',
} as const;