# lib/api-zod/src/generated/types/chartPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXA Payment Crypto API
 * OpenAPI spec version: 0.1.0
 */

export interface ChartPoint {
  timestamp: string;
  price: number;
}